Update for latest available PKHeX Core.

+++ SysBot.Pokemon.Discord/Helpers/AutoLegalityExtensionsDiscord.cs
@@ -48,7 +48,7 @@ namespace SysBot.Pokemon.Discord
         {
             content = ReusableActions.StripCodeBlock(content);
             var set = new ShowdownSet(content);
-            var sav = AutoLegalityWrapper.GetTrainerInfo(gen);
+            var sav = AutoLegalityWrapper.GetTrainerInfo((byte)gen);
             await channel.ReplyWithLegalizedSetAsync(sav, set).ConfigureAwait(false);
         }
